local storage working for project and notes
need to look at issue with task cards not saving properly
need to strip it back to basics and see if it is a problem with the save/load or the task card itself.
most likely issue is the way the taskcard is built so storing is becaoming an issue.

best to remove features from taskcard and see if we can get it to save properly before adding them back in.